Allow for the RTP clock rate to differ from the actual sampling rate
RFC 3551 requires that some codecs (namely G.722 as far as we are
concerned) be sent at an RTP clock rate that differs from the actual
sampling rate.

In particular, this means that the number of samples accounted for in
the RTP timestamp may differ from the actual number of samples that were
captured or played.